requirements.txt
django
newsapi-python
uvicorn
aiohttp
asyncio
requests
nest_asyncio
openai
gunicorn==20.0.4
Pillow==9.1.0
psycopg2-binary==2.9.1